Changed the anchor tag on index.html, added e new anchor tag in kalojam.html
On index.html, the anchor tag had a attribute which would cause it to open the new page in a separate tab. Removed this, so new pages would open in the current tab.

On kalojam.html, added a new anchor tag, that links back to the main homepage [index.html]